Moved CircleCI Robo to a BLT command (#318)
* Moved CircleCI Robo to a BLT command
